Refactor out into helper
